Move folders to root
- Libraries/Partout-Framework -> .
- Libraries/Partout-Source -> src
- . -> bin
