+++ .github/workflows/test.yml
@@ -25,7 +25,7 @@ jobs:
           submodules: false
       - name: Set development environment
         run: |
-          sed -i '' "s/environment = .production/environment = .onlineDevelopment/" "Libraries/Partout-Framework/Package.swift"
+          sed -i '' "s/environment = .production/environment = .onlineDevelopment/" "Framework/Package.swift"
       - name: Run tests
         run: |
           swift test
