+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "Libraries/Partout-Source"]
-	path = Libraries/Partout-Source
+[submodule "src"]
+	path = src
 	url = git@github.com:passepartoutvpn/partout-source.git
